requirements.txt
Django==4.2.16
django-model-utils==4.5.1
gunicorn==22.0.0
django-debug-toolbar==4.4.6
django-jazzmin==3.0.0
requests==2.32.3
click~=8.1.7
cookiecutter==2.6.0
GitPython==3.1.43

setuptools==70.3.0
djangorestframework==3.15.2
